+++ docs/release_notes.rst
@@ -11,6 +11,7 @@ Changes from 0.8.0 to 0.9.0
 
   - PRISM products (``SW_OPER_MITx_LP_2F``, ``SW_OPER_MITxTEC_2F``, ``SW_OPER_PPIxFAC_2F``)
   - Multi-mission magnetic products (``CS_OPER_MAG``, ``GRACE_x_MAG``, ``GFx_OPER_FGM_ACAL_CORR``)
+  - Swarm spacecraft positions (``SW_OPER_MODx_SC_1B``)
 
 - Fixed missing auxiliary "dDst"
 - Fixed fetching longer time series of hourly observatory products